Expand the scope of ignored PHP-Doc tags (#1516)
* Expand the scope off ignored PHP-Doc tags

Closes: #1399

Expand the list of tags that produce placeholder names that should be ignored by Deptrac from `@template` to also include `@template-covariant`, `@phpstan-type` and `@phpstan-import-type`.

This is coincidentally only necessary for Nikic Parser and not PHPStan parser. That does it out of the box. Also, this is a whack-a-mole, as the list of these types of tags can expand in the future.

* SA fixes